requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cssbeautifier==1.15.1
cssselect2==0.7.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.7
django-debug-toolbar==4.4.6
django-filer==3.1.4
django-polymorphic==3.1.0
djlint==1.34.1
docopt==0.6.2
easy-thumbnails==2.9
EditorConfig==0.12.4
executing==2.0.1
fastjsonschema==2.20.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.12.25
reportlab==4.2.2
requests==2.32.3
rpds-py==0.19.1
setuptools==72.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
